[URP][XR]Add support to generate Skybox shaders form ShaderGraph
This solves an issue where the skybox would be depth clipped in stereo single-pass because the constant buffer for the matrices would not contain the right matrices...
The Skybox pass currently use the built-in renderer, so we need to make sure that the generated shader will user the built-in stereo matrices constant buffer instead of the one from URP.

To do that, ShaderGraph will define USING_BUILTIN_STEREO_MATRICES for shaders built with the "Skybox" material type. This will handle all the necessary changes to make built-in stereo matrices work with URP and use them for single-pass stereo shader variation.